requirements.txt
numpy
torchio
albumentations
tensorboard
h5py
matplotlib
seaborn
spectral
scikit-learn==0.22.2
torchsummary
torchvision==0.5.1